+++ headers/data/overlay29/move_effects.h
@@ -1,0 +1,6 @@
+#ifndef HEADERS_DATA_OVERLAY29_MOVE_EFFECTS_H_
+#define HEADERS_DATA_OVERLAY29_MOVE_EFFECTS_H_
+
+extern int32_t LUNAR_DANCE_PP_RESTORATION;
+
+#endif
